+++ examples/iced/lib/iced_example/mailer.ex
@@ -1,0 +1,3 @@
+defmodule IcedExample.Mailer do
+  use Swoosh.Mailer, otp_app: :iced_example
+end