+++ examples/iced/lib/iced_example_web/controllers/page_controller.ex
@@ -1,0 +1,7 @@
+defmodule IcedExampleWeb.PageController do
+  use IcedExampleWeb, :controller
+
+  def home(conn, _params) do
+    render(conn, :home)
+  end
+end