+++ examples/iced/test/iced_example_web/controllers/error_json_test.exs
@@ -1,0 +1,12 @@
+defmodule IcedExampleWeb.ErrorJSONTest do
+  use IcedExampleWeb.ConnCase, async: true
+
+  test "renders 404" do
+    assert IcedExampleWeb.ErrorJSON.render("404.json", %{}) == %{errors: %{detail: "Not Found"}}
+  end
+
+  test "renders 500" do
+    assert IcedExampleWeb.ErrorJSON.render("500.json", %{}) ==
+             %{errors: %{detail: "Internal Server Error"}}
+  end
+end